Fix incorrect modify removed item conflict detection
This commit fixes a bug in conflict and update detection where a modify
removed item conflict was found instead of a remove edge update. This
occurred because the lamport clock used in the comparison was the same.

Interestingly, this bug was not picked up in unit tests nor do unit
tests performing equivalent logic to the new integration test catch the
change. Thus, a new integration test that catches the bug has been
added.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Zachary Hamm <zack@systeminit.com>