chore(db): add pgbouncer to our dev setup
This PR adds pgbouncer to our dev setup. It moves the "real" postgres
instances to 7432 (si) and 8432 (si-test) respectively, replacing them
with two different pgbouncers. This change is primarily to ensure that
we avoid any threat of pool exhaustion in the test suite, but has the
added benefit of making the dev stack closer to our production
deployment (which would also include pgbouncer.)

PgBouncer configures a pool per hostname+database. Since each of our
dal integration tests creates its own database, it thereby also creates
its own pool. To manage this, we create a pool size of 10 per
hostname+database tuple, which we will then reap after 10 idle seconds.
The result is that the current test suite peaks somewhere around 300
tests, but very rapidly shrinks. We should have plenty of headroom.

The only tricky thing here is that we alter the migration paths for
tests, as we migrate with the underlying database connection directly.
When we do, we must forcibly disconnect the bouncer - so if you happen
to be logged in to postgres, get ready to be logged out when we need to
make the original template database for tests.

Signed-off-by: Adam Jacob <adam@systeminit.com>