Restore all of func authoring
This is it. This commit restores all of func authoring. The features
affected include, but are not limited to the following: test func
execution, save and exec func, save func fixes, attribute prototype
argument associations fixes, delete func and more. Let's talk about it.

All routes related to func authoring return with this commit. You can
now delete funcs, perform test executions on them, and save and exec
funcs. Among these restored features comes fixes to how associations
are handled, general save func fixes and other quality of life
improvements. This also includes the next pass at evaluating and
generating input types for `Props`, leaf nodes and `Components`.

There _will_ be bugs. Please see the disclaimer section for more
information.

Primary changes:
- Restore `save_and_exec` func with `save_func` and `execute` (func) as
  the backing functions
- Restore `exec_func` with `dummy_execute_func` as the backing function
- Restore `list_input_sources` used for attribute bindings as their own
  sub-module
- Restore `delete_func`
- Fix input types for leaf functions
- Polish edges and existing bits of remaining func authoring logic (i.e.
  restored `WsEvents`)
- Remove unused modules, including but not limited to the following:
  attribute view, prop tree, reconciliation prototype and index map

Secondary changes:
- Add ability to list ordered child props
- Ensure ordering of props at a single "level" is stable for the
  property editor
- Add ability for props to know their input types via `ts_types` (the
  implementation is not fully correct because type collection is
  inverted, but it is close)
- Remove unused modules, including but not limited to the following:
  attribute view, prop tree, reconciliation prototype and index map

Known issues:
- Action prototypes via `save_func` are re-created when they shouldn't
  be and we should be handling this logic in a different way
- Auth prototypes via `save_func` are re-created when they shouldn't
  be and we should be handling this logic in a different way
- Input source and attribute binding parameter collection and display
  functionality is not polished
- Creating or updating attribute bindings may not function as intended
- `AttributePrototypeArgumentBags` need to be able to specify a `PropId`
  (replaces the implicit `InternalProviderId` field from the old engine)
- Reloading after `save_func` or updating attachments sometimes does not
  persist

Due to the scope of this change, it is likely that features contain rough
edges or missing elements (especially test execution and "save_and_exec").
The goal of this commit is to unblock future work related to func authoring.

Usually, I'd aim for more certainty with the quality of my commits, but the
opportunity cost of holding up individual piecemeal authoring features is
greater than the cost of the team swarming and polishing the remaining bits
and bugs. Moreover, the func authoring experience is relatively sandboxed,
so the rest of the system should not regress if the func authoring
experience is untouched.

Signed-off-by: Nick Gerace <nick@systeminit.com>