feat(rebaser): re-impl server w/activity stream & per-change set queues
This is a rather large change which re-implements the Rebaser server
code using several new architectural patterns and frameworks.

The Rebaser server now processes a filtered activity stream (provided by
the LayerDb infrastructure) of rebaser requests and for each message
checks to see if there is a corresponding running task for the
associated change set. If not a new "micro-service"/"actor" is spawned
which consumes requests for the given change set from a NATS Jetstream
work-queue in a one-at-a-time fashion. The work-queue can be shared
across multiple running instances of Rebaser servers, meaning that a
group of Rebasers can share the coordinated processing of all messages
while simultaneously allowing work across multiple change sets to run
concurrently.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>