feat(dal): add exclusive outgoing edge conflicts to rebase
Many nodes assume complete control of their direct, outgoing neighbors,
and allowing them to be silently merged with other neighbors from
another changeset would create an incorrect graph. For example, an
AttributeValue should only ever have one outgoing Prototype edge
defining the "component specific" prototype for that value. If another
user merges in a changeset with an updated prototype edge that we have
never seen in our change set, that should be considered a conflict,
since otherwise the result of the rebase would be an AttributeValue with
*two* Prototype edges, an incorrect state. The same applies for
"Contains" edges and many others. This change will produce an
`ExclusiveEdgeMismatch` conflict when the snapshot being rebased (the
`to_rebase`) has a container with edges not seen by the `onto`, *if*
that container defines the edge kinds as exclusive edges. Be sure to add
these exclusions the node weight when adding a new edge kind, if it
makes sense to do so.