Polish func authoring: attribute bindings, test executions, and more
This commit is a follow-up to 77eeda3cf78947c5534c6a379515328159007f85
and polishes a lot of the func authoring experience.

Attribute Bindings Changes:

Updating and adding attribute bindings for components has been
polished. Adding, editing and removing attribute bindings for components
should work as intended. However, there are remaining issues with
attribute bindings for schema variants, but they will be touched on
later in the commit.

There are now an integration test for creating a new prototype and
prototype argument for an existing attribute func with an existing
attribute binding.

FuncAssociations Changes:

A key change to getting attribute bindings in a better state as well as
overall improvements was to restore the idea of "implicit internal
providers" for attribute func associations. Essentially, a "prop_id"
field now takes a similar role to before. The collection and mutation of
the field is now backed by new and old integration tests alike.

Func Test Execution Changes:

All naming related to the "test execution" of funcs has been realigned.
The "critical" section layer of func execution has been sandboxed in
order to avoid transaction rollback scenarios (i.e. no bindings, return
values nor execution recorded in old database tables).

Func Deletion and Detachment Changes:

Detaching funcs and deleting funcs works. Func deletion will be
unsuccessful if there are existing associations, but we do not make it
clear in the UI when this happens.

Misc Changes:

- Tune down log level for checking the rebaser response payload

Known Issues:

- Working with attribute bindings at the schema variant level seens an
  "attribute prototype has been orphaned" error
- Attribute funcs corresponding to attribute bindings at the component
  level do not appear to run on apply
- Editing an attribute func in the interfere does not show existing
  options in the modal
- Static argument values are not available in the modal for selection
- It is not obvious when func associations existing blocks func deletion

Signed-off-by: Nick Gerace <nick@systeminit.com>