Remove badge for nightly CI builds
This commit removes the badge for nightly CI builds. With the new
engine, we believe we do not need nightly CI runs (for now) because
equivalent tests are expected to be much faster to run. Combine that
with the fact that nightly runs have not been ran in nearly two months
and it's evident that it's time to retire the badge.

Signed-off-by: Nick Gerace <nick@systeminit.com>