Cleanup references to disallowing editing of builtin funcs
This commit cleans up references to disallowing editing of builtin
funcs. It is not yet decided what the user can and cannot edit in the
new engine.

In addition, dead code is removed from sdf tests and the dal, and
we will no longer error on saving a func with a kind that does not have
associations.

Signed-off-by: Nick Gerace <nick@systeminit.com>