Fix test executions for attribute funcs and missing payload value
This commit fixes test executions for attribute funcs. It also scopes
test execution abilities to solely attribute, code generation and
qualification funcs.

The existing code assumes that there was only one binding for attribute
funcs. Attribute funcs may have multiple bindings, which means that they
can have different input sources (AttributePrototypeArguments) for the
same FuncArguments. For example: if you have a FuncArgument that takes
in a string called "toddhoward", it is possible for one binding to get
that value from "/root/si/name" and another binding to get that value
from an input socket who provides a string.

The existing code also assumed that there was only one FuncArgument.
Now, we derive the arguments from the backend and avoid walking the
result of "component/json" (as well as assuming that the FuncArgument
name will match the name of a given Prop).

This commit also fixes a bug where "/root/resource/payload" would not be
returned from the "component/json" route due to a dropped field during
de/serialization.

Additional changes:
- Placeholder text colors for VormInput now work in both light and dark
  mode
- Split selector details out into a new component: FuncTestSelector
- Split argument preparation for prototype execution into its own method
  with a corresponding integration test
- Remove unused code related to fetching latest func executions
- Add new "/attribute" endpoint and alongside a route to provide
  prototype arguments

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Wendy Bujalski <wendy@systeminit.com>