Updating encrypted contents triggers dependent values update
This commit ensures that updating the encrypted contents of a secret
results in dependent values update being triggered.

For context, in the frontend, the encrypted contents for secrets are
updated centrally, which means that if multiple attribute values "use"
that secret, they must be processed in dependent values update. This
presents us with a few design challenges unique to the new engine, but
none that are/were unsolveable.

The first architectural challenge was to make an implicit relationship
(i.e. setting SecretId as a string on a prop with a "secret" widget
kind) into an explicit one. The solution: add a new value source for
AttributePrototypeArguments. The new value source indicates how the sole
argument should be assembled when executing from prototype func (i.e
"si:identity"). With this change, we now have an explicit relationship
linking the attribute subsystem to the secret.

A notable result of the aforementioned solution is that the encrypted
secrety key is now passed through the attribute subsystem rather than
the SecretId alone. We'll touch back on this when talking about the
third architectural challenge.

The second architectural challenge was to allow SecretIds to be passed
into dependent values update. Part of the long term vision for dependent
values update is to be able to take any node on a
WorkspaceSnapshotGraph, determine what depends on it and execute
relevant functions. Working with Jacob Helwig (@jhelwig) and Zachary
Hamm (@zacharyhamm) led to the solution in this commit: ensure the
dependency graph can take in arbitrary values, but continue using
attribute values for the remainder of dependent values update. There is
a little blast radius here given the proximity to ComputeValidations and
job de-duplication code, but the bulk of the changes are mostly
self-contained.

The third architectural challenge was to inform the frontend (via the
property editor) of which secret had been selected for a given
property. The solution: send the SecretId along with the encrypted
secret key in the attribute subsystem. In the frontend, the list of
secrets available purely uses the SecretIds, which are stable. We could
have used solely the key and provided a way to find the secret via the
key. We also could have sent up the entire object when assembling the
property editor values. However, the implementation chosen involves
scrubbing the encrypted secret key from the property editor values. This
reduce friction in the frontend since "app/web" is solely concerned with
SecretIds and not the encrypted secret keys themselves.

There are a lot more changes in this commit, including moving the
"before_funcs" module underneath the "secrets" module, but that's the
breakdown of the primary objective and implementation.

Misc Changes:
- Move "before_funcs" module underneath the "secrets" module
- Expand ValueSource display implementation
- Rename "Visibility::new_head" to "Visibility::new_head_fake" to
  indicate intent alongside a doc comment
- Allow DependentValuesUpdate to take any node as input, though only
  contents nodes for Secrets and AttributeValues are evaluated at
  present

Signed-off-by: Nick Gerace <nick@systeminit.com>