chore(dal-test): disable OpenTelemetry in integration tests
The OpenTelemetry setup in our integration tests are close to operating,
while still having issues when executing. The largest problem to
overcome is dealing with failing/panicking tests that don't easily
provide a way to the OpenTelemetry collector to flush and clean up
appropriately.

Future work may try to track tests in flight and infer when the last
test is done so as to wait for a clean shutdown of OpenTelemetry, but
this is very difficult considering there is no low level configuration
of the Rust test suite harness. In other words, it'll be something of a
hack in the end...

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>