Add metrics to pinga jobs
Improves telemetry/metrics of `execute_job_task`.

Logs the number of concurrent tasks and limit every time a new job is
spawned in pinga. Also logs if pinga is at full capacity, which means
no new tasks will be consumed until one is finished.

We also are logging the status of execution, jobs without status are
still being executed, so we can easily keep track of what is going on
at each instance.