Fix func attachment to assets and polish authoring client
This commit fixes the ability to attach funcs to assets. In the
frontend, we needed to ensure that we were using SchemaVariantId
corresponding to the default SchemaVariant of the selected asset. In the
backend, we needed to use "WsEvent::FuncSaved" rather than a re-fetched
FuncView in order to ensure that associations in the frontend are up to
date.

This commit also polishes the func authoring client. Tracing has been
added and tuned for all func authoring client methods (descending until
"work" layer functions). In addition, re-exported functions have been
replaced with their inner contents. Now, re-exported functions merely
provide additional functionality needed in order to not bloat the
"authoring.rs" file.

Updating associations for action and authentication funcs has been
polished with the ability to cleanup undesired prototypes that may
linger.

Additional fixes in this commit include, but are not limited to
improving reactivity when deleting funcs, improving reactivity when
detaching funcs and adding additional error checks when updating
associations.

Summarized Changes:
- Use "WsEvent::FuncSaved" instead of re-feteched func views for save
  func routes
- Use the default SchemaVariantId via the selected asset for the attach
  func modal
- Streamline and extend tracing for func authoring client methods (as
  well as some inner functions), including only skipping "ctx" rather
  than all fields
- Take the contents of re-exported functions for the authoring client
  and use the sub-modules to provide additional functionality in order
  to not bloat "authoring.rs"
- Clean up undesired prototypes for action and authentication funcs
- Catch errors when the associations kind and func kind do not match
- Fix "WsEvent::func_saved" by using the correct payload name
- Fix asset func list not re-loading when detaching a func
- Fix open tab error when deleting a func

Signed-off-by: Nick Gerace <nick@systeminit.com>