Fix secret attribute value containing an object instead of a string
This commit fixes the attribute values corresponding to individual
secrets containing an object instead of a string.

The encrypted secret key is now passed through the attribute subsystem
alone. Since the property editor still needs to reference which
encrypted secret key belongs to which secret, we determine the origin
secret dynamically when assembling the property editor (~3.8ms wall
clock time for 1000 secrets with a 3970X system on Pop!_OS 22.04 LTS
with Linux kernel 6.8.0).

As a result of these changes, we were able to remove intermediary types
with a lot of footguns as well as move the encrypted secret key to the
new SecretNodeWeight kind. Overall performance and consistency in
results should be net positive.

Other Changes:

Another change this commit brings is the inclusion of our first
integration test... "bench". We want to ensure that looking up which
secret an encrypted secret key is derived from is fast, so the new test
will fail if the function takes too long. We may want to use this logic
to create tests where similar performance scrutiny is needed.

Finally, we are now using the "create_xxhash_type" macro to ensure that
the EncryptedSecretKey works with "si-layer-cache".

Signed-off-by: Nick Gerace <nick@systeminit.com>