Dispatch Actions V2 after successful rebase of default change set for workspace
Whenever we have successfully rebased the default change set for a
workspace, we need to dispatch any eligible Actions (and mark them as
dispatched). By checking which Actions are eligible to dispatch on
successful rebase, we ensure that any changes to the default change set
(typically `HEAD`, or `main`) will automatically cause relevant Actions
to run, such as an Action being removed from the graph because it
succeeded, or new Actions being queued from merging a change set into
the default change set.