Fix attaching two auth or action funcs to the same variant
This commit contains two fixes that present themselves as the same
issue: attaching auth or action funcs to the same variant silently
fails. The individual fixes are related in logic changes, but are
different overall. Essentially, we were not handling the "existing vs.
removed" SchemaVariantId logic for the associations properly. Now, we
ensure that existing associations are not accidentally dropped.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Paul Stack <paul@systeminit.com>