Fix detach for attribute and code gen
This commit fixes detaching for attribute and code generation funcs.

For attribute funcs, we needed fallback to removing the prototype at
the schema variant level if an attribute prototype could not be found
at the component level (i.e. the prototype did not correspond to an
attribute value). There is now an integration test tracking this
behavior.

For code generation funcs, we needed to ensure that we were using the
detach ref. No new integration tests were added for this behavior
because the bug was purely a frontend usage bug.

BUG-185
BUG-205

Signed-off-by: Nick Gerace <nick@systeminit.com>