fix(dal): recalculate merkle tree hash in import_subgraph
If import_subgraph cannot find an equivalent node in to_rebase, it will
add the node from onto directly to the graph, and then call update_edge
to add the edges. However, update_edge does not recalculate the merkle
tree hash, and so the node added via add_node will have the merkle tree
hash of node with no children. Add a call to update_merkle_tree_hash at
the end of import subgraph to ensure the hash is updated properly.