Do not dispatch Actions with pending dependent value updates
While there are AttributeValues for any given Component that are still
involved in a DependentValuesUpdate, we want to hold off on dispatching
any Actions for those Components. This will let us make sure that the
Action will be operating with the most up-to-date data possible in the
model.

Now that we're tracking the root nodes for the DependentValuesUpdate
graph of AttributeValues to update, we can directly check the Components
that are involved in the pending DependentValuesUpdate to see if an
Action would operate on an involved Component.