Include node ID and kind in Conflict & Update records
Previously, we were only including the (internal to the graph)
`NodeIndex` in the structs for representing conflicts & updates. While
this is all that's strictly necessary for performing the updates and
(eventually) having SDF report what the conflicts are, having more
contextual information such as the node weight kind, and the ID of the
node makes things a bit easier for debugging issues when looking at
spans & traces.