fix(dal): add recently seen clocks to edge weights
EdgeWeights had no "recently seen" vector clocks. This lead to
situations where we produced a conflict but should not have, because we
only had the "first seen" clock to work with. This adds the recently
seen vector clock to the graph. However, this cannot be done without a
new version of the WorkspaceSnapshotGraph, so migration logic has been
implemented to automatically migrate the snapshot on first read.