Move check for Conflict::ExclusiveEdgeMismatch to operate on the net changes for the outgoing edges
Previously, the check was not taking into account that while there may
be an `Update::NewEdge` that is creating an outgoing edge of the same
`EdgeWeightKind` as an existing edge, there might also be an
`Update::RemoveEdge` that is removing that pre-existing edge.

We now check to make sure that after all additions & removals of
outgoing edges for a specific `EdgeWeightKind`, the net result is that
we have no more than a single edge.

The following exclusive edge mismatch conditions have been removed as
they are more nuanced than "Node X can have at most 1 of edge kind Y",
and will be addressed at a later date.

  * `AttributeValue` - `Contain`: `AttributeValue`s can have multiple
    outgoing `Contain` edges. The restriction on whether or not a
    specific `Contain` edge is a semantic conflict with another depends
    on information about the `Prop` tree structure, and looking at the
    `key` portion of the `Contain` edge weight.

  * `Component` - `FrameContains`: The exclusivity of the
    `FrameContains` edge is that a `Component` can only have one
    *incoming* `FrameContains` edge. An arbitrary number of *outgoing*
    `FrameContains` are allowed.

  * `Func` - `Use`: A `Func` can have multiple outgoing `Use` edges, and
    is a requirement for any `Func` that has multiple `FuncArgument`s.

  * `Prop` - `Prototype`: As with `AttributeValue` and `Contain`, this
    requires looking at the prop tree structure, and taking into
    consideration the `key` portion of the `Prototype` edge weight.

  * `Prop` - `Use`: As with `FrameContains`, the direction of the
    exclusivity is on *incoming* edges, not on the number of *outgoing*
    edges.