Handle Some(serde_json::Value::Null) more gracefully when setting nested values
Rather than trying to use `si:setString`, `si:setBool`, or any similar
function with a `serde_json::Value::Null`, we now translate this to
behave as though we had gotten a `None`, and use `si:unset`.

Ideally we would handle things in such a way that we can differentiate
between "there is no value here" (`None`), and "the value here is
`null`" (`Some(serde_json::Value::Null)`), but this is hard enough to do
within JavaScript/JSON on its own, let alone round tripping data in &
out of JSON multiple times before dealing with it.