Detect ChangeStatus::Added for Components in summary diagram
Rather than reporting that all Components are `ChangeStatus::Added`, we
now detect whether the Component would be added to the base change set,
if we were to merge this change set into it. If we are looking at the
default change set for a Workspace, then we consider it to not have a
base change set, even though it is possible to have one defined.

This does not detect `ChangeStatus::Modified`, or the `ChangeStatus` for
Edges yet.