fix(dal): cleanup vector clock entries except node writes
All *node* write vector clock entries have to be preserved for conflict
detection to work correctly. However, the "seen" clocks on nodes only
need changeset clocks, and the write/seen clocks on edges also only need
changeset clocks. (Edges are never mutated in our system, only created
or destroyed, so the write clock is never updated and conflict detection
only happens between the to_rebase and onto clocks specifically, which
are either a changeset clock or the editing clock for that edit
session). The majority of our clock entries are "seen" clocks, not write
clocks, so this will work to keep the snapshot size small.